remove CEM and add permission framework

+++ src/main/java/de/mendelson/util/security/keydata/KeydataAccessDB.java
@@ -81,11 +81,12 @@ public class KeydataAccessDB {
     /**
      * Returns the timestamp in ms the keydata entry has been modified last
      */
-    public long getLastChanged(int purpose) {
+    public long getLastChanged(int purpose, int userId) {
         try (Connection configConnectionAutoCommit = this.dbDriverManager.getConnectionWithoutErrorHandling(IDBDriverManager.DB_CONFIG)) {
             try (PreparedStatement statement = configConnectionAutoCommit.prepareStatement(
-                    "SELECT lastchanged FROM keydata WHERE purpose=?")) {
+                    "SELECT lastchanged FROM keydata WHERE purpose=? AND user_id=?")) {
                 statement.setInt(1, purpose);
+                statement.setInt(2, userId);
                 try (ResultSet result = statement.executeQuery()) {
                     if (result.next()) {
                         long timestamp = result.getLong("lastchanged");
@@ -100,14 +101,17 @@ public class KeydataAccessDB {
     }
 
     /**
-     * Returns the key storage of the requested purpose
+     * Returns the key storage of the requested purpose and user
+     * @param purpose KEYSTORE_USAGE_TLS or KEYSTORE_USAGE_ENC_SIGN
+     * @param userId User ID (0 = admin/system, >0 = specific user)
      */
-    public KeystoreData getKeydata(int purpose) {
+    public KeystoreData getKeydata(int purpose, int userId) {
         try (Connection configConnectionAutoCommit = this.dbDriverManager
                 .getConnectionWithoutErrorHandling(IDBDriverManager.DB_CONFIG)) {
             try (PreparedStatement selectStatement = configConnectionAutoCommit.prepareStatement(
-                    "SELECT * FROM keydata WHERE purpose=?")) {
+                    "SELECT * FROM keydata WHERE purpose=? AND user_id=?")) {
                 selectStatement.setInt(1, purpose);
+                selectStatement.setInt(2, userId);
                 try (ResultSet result = selectStatement.executeQuery()) {
                     if (result.next()) {
                         byte[] keyData = this.dbDriverManager.readBytesStoredAsJavaObject(result, "storagedata");
@@ -157,7 +161,7 @@ public class KeydataAccessDB {
      * Inserts a new key store into the database if none exists so far
      */
     public void insertKeydataFromFileIfItDoesNotExistInDB(Logger logger, Path keystoreFile,
-            String storageType, int purpose, String securityProvider) {
+            String storageType, int purpose, String securityProvider, int userId) {
         String transactionName = "Keydata_insert";
         try (Connection configConnectionNoAutoCommit = this.dbDriverManager
                 .getConnectionWithoutErrorHandling(IDBDriverManager.DB_CONFIG)) {
@@ -170,8 +174,9 @@ public class KeydataAccessDB {
                             "keydata"});
                 boolean entryExists = true;
                 try (PreparedStatement checkStatement = configConnectionNoAutoCommit.prepareStatement(
-                        "SELECT COUNT(1) AS counter FROM keydata WHERE purpose=?")) {
+                        "SELECT COUNT(1) AS counter FROM keydata WHERE purpose=? AND user_id=?")) {
                     checkStatement.setInt(1, purpose);
+                    checkStatement.setInt(2, userId);
                     try (ResultSet result = checkStatement.executeQuery()) {
                         if (result.next()) {
                             entryExists = result.getInt("counter") > 0;
@@ -180,13 +185,14 @@ public class KeydataAccessDB {
                     if (!entryExists) {
                         byte[] keystoreData = Files.readAllBytes(keystoreFile);
                         try (PreparedStatement insertStatement = configConnectionNoAutoCommit.prepareStatement(
-                                "INSERT INTO keydata(storagedata,storagetype,purpose,lastchanged,securityprovider)"
-                                + "VALUES(?,?,?,?,?)")) {
-                            this.dbDriverManager.setBytesParameterAsJavaObject(insertStatement, 1, keystoreData);
-                            insertStatement.setInt(2, keystoreTypeStrToInt(storageType));
-                            insertStatement.setInt(3, purpose);
-                            insertStatement.setLong(4, System.currentTimeMillis());
-                            insertStatement.setString(5, securityProvider);
+                                "INSERT INTO keydata(user_id,storagedata,storagetype,purpose,lastchanged,securityprovider)"
+                                + "VALUES(?,?,?,?,?,?)")) {
+                            insertStatement.setInt(1, userId);
+                            this.dbDriverManager.setBytesParameterAsJavaObject(insertStatement, 2, keystoreData);
+                            insertStatement.setInt(3, keystoreTypeStrToInt(storageType));
+                            insertStatement.setInt(4, purpose);
+                            insertStatement.setLong(5, System.currentTimeMillis());
+                            insertStatement.setString(6, securityProvider);
                             insertStatement.executeUpdate();
                         }
                     }
@@ -210,7 +216,7 @@ public class KeydataAccessDB {
      * purpose that is already in the system. Else the system data will left
      * untouched.
      */
-    public void updateKeydata(byte[] keystoreData, String keystoreType, int purpose, String securityProvider) {
+    public void updateKeydata(byte[] keystoreData, String keystoreType, int purpose, String securityProvider, int userId) {
         String transactionName = "update_Keydata";
         try (Connection configConnectionNoAutoCommit = this.dbDriverManager
                 .getConnectionWithoutErrorHandling(IDBDriverManager.DB_CONFIG)) {
@@ -224,13 +230,14 @@ public class KeydataAccessDB {
                 try (PreparedStatement updateStatement = configConnectionNoAutoCommit.prepareStatement(
                         "UPDATE keydata SET "
                         + "storagedata=?,storagetype=?,lastchanged=?,securityprovider=? "
-                        + "WHERE purpose=?")) {
+                        + "WHERE purpose=? AND user_id=?")) {
                     this.dbDriverManager.setBytesParameterAsJavaObject(updateStatement, 1, keystoreData);
                     updateStatement.setInt(2, keystoreTypeStrToInt(keystoreType));
                     updateStatement.setLong(3, System.currentTimeMillis());
                     updateStatement.setString(4, securityProvider);
                     //condition
                     updateStatement.setInt(5, purpose);
+                    updateStatement.setInt(6, userId);
                     updateStatement.executeUpdate();
                     this.dbDriverManager.commitTransaction(transactionStatement, transactionName);
                 } catch (Throwable e) {
@@ -246,7 +253,7 @@ public class KeydataAccessDB {
     /**
      * Deletes the keydata entry of a special purpose
      */
-    public void deleteKeydata(int purpose) {
+    public void deleteKeydata(int purpose, int userId) {
         String transactionName = "delete_Keydata";
         try (Connection configConnectionNoAutoCommit = this.dbDriverManager
                 .getConnectionWithoutErrorHandling(IDBDriverManager.DB_CONFIG)) {
@@ -258,8 +265,9 @@ public class KeydataAccessDB {
                         new String[]{
                             "keydata"});
                 try (PreparedStatement deleteStatement = configConnectionNoAutoCommit.prepareStatement(
-                        "DELETE FROM keydata WHERE purpose=?")) {
+                        "DELETE FROM keydata WHERE purpose=? AND user_id=?")) {
                     deleteStatement.setInt(1, purpose);
+                    deleteStatement.setInt(2, userId);
                     deleteStatement.executeUpdate();
                     this.dbDriverManager.commitTransaction(transactionStatement, transactionName);
                 } catch (Throwable e) {
